+++ micromanager_gui/_gui/explore_sample.ui
