+++ micromanager_gui/_gui/mm_snap_and_tabs.ui
