+++ micromanager_gui/_gui/multid_gui.ui
